requirements.txt
flask==3.1.3
google-genai==1.47.0
python-dotenv==1.2.1
